# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Management System API",
        "description": "RESTful API для системы управления блогом с аутентификацией и системой ролей",
        "contact": {
            "email": "u.azamat.1997@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Основной сервер API"
        }
    ],
    "paths": {
...
        "/api/blog/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить список постов",
                "operationId": "c9d1093f043f8b7c8e657fb31a37a2c4",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтрация по статусу (draft/published)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Фильтрация по ID автора",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Фильтрация по slug категории",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Фильтрация по slug тега",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по заголовку и содержимому",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Сортировка (например: created_at)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Направление сортировки (asc/desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Создание нового поста",
                "operationId": "b775b4fbac8062f327b6bf10e4cbbb06",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Заголовок поста"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Содержимое поста"
                                    },
                                    "excerpt": {
                                        "type": "string",
                                        "example": "Краткое описание"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "featured_image": {
                                        "description": "Изображение обложки поста",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Пост успешно создан"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/blog/posts/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить конкретный пост",
                "description": "Возвращает один пост по ID",
                "operationId": "3d09fc7f2f3e86c2f92fb445c2eb512e",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Пост не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Author": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Azamat"
                    },
                    "email": {
                        "type": "string",
                        "example": "azamat@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Технологии"
                    },
                    "slug": {
                        "type": "string",
                        "example": "texnologii"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание категории"
                    },
                    "color": {
                        "type": "string",
                        "example": "#3B82F6"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "PHP"
                    },
                    "slug": {
                        "type": "string",
                        "example": "php"
                    },
                    "color": {
                        "type": "string",
                        "example": "#777BB4"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Пост с изображением от Пользователя"
                    },
                    "slug": {
                        "type": "string",
                        "example": "Post-s-izobrazeniem-ot-polzovatelya"
                    },
                    "excerpt": {
                        "type": "string",
                        "example": "Краткое описание поста"
                    },
                    "content": {
                        "type": "string",
                        "example": "Контент поста"
                    },
                    "featured_image": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/images/posts/example.jpg"
                    },
                    "thumbnail": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/images/posts/example.jpg"
                    },
                    "status": {
                        "type": "string",
                        "example": "published"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "views_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}